requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-storage-blob
azure-core>=1.0.0
certifi==2024.6.2
cffi==1.16.0
distro==1.9.0
exceptiongroup==1.2.1
fastapi
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.35.3
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv
scipy==1.13.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn
wsproto==1.2.0

# Google Calendar API
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Scheduling
schedule>=1.2.0

# Local Whisper for transcription
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# LangChain for Azure OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5

# Local LLM Support (Ollama, LLaMA.cpp)
ollama>=0.6.0
llama-cpp-python>=0.3.0
langchain-community>=0.4.0

# Additional dependencies for Whisper
ffmpeg-python>=0.2.0

# CLI tool
click>=8.1.0
rich>=13.0.0